* master:
  typos
  udpated notification doc
  typos
  updated docs
  Casting age and value
  Integer casts
  Made NSInteger casts
  Casted macro on testVersionGetters
  Removed duplicate RLMRealm accessor method from RLMTestCase

Conflicts:
	src/realm/objc/test/group.m
	src/realm/objc/test/group_misc_2.m
	src/realm/objc/test/tutorial.m

+++ src/realm/objc/test/version.m
@@ -10,9 +10,9 @@
 
 -(void)testVersionGetters
 {
-    XCTAssertEqual(REALM_VERSION_MAJOR, [RLMVersion major] );
-    XCTAssertEqual(REALM_VERSION_MINOR, [RLMVersion minor] );
-    XCTAssertEqual(REALM_VERSION_PATCH, [RLMVersion major] );
+    XCTAssertEqual((NSInteger)REALM_VERSION_MAJOR, [RLMVersion major] );
+    XCTAssertEqual((NSInteger)REALM_VERSION_MINOR, [RLMVersion minor] );
+    XCTAssertEqual((NSInteger)REALM_VERSION_PATCH, [RLMVersion major] );
 
     NSString *ver1 = [RLMVersion version];
     NSString *ver2 = [NSString stringWithFormat:@"%d.%d.%d",
